pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "terminal-bridge"
dynamic = ["version"]
description = "Model-agnostic remote Mac terminal access for AI agents. One command to install, works with Cursor, Ollama, Claude CLI, and any AI tool."
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Rajesh Rout", email = "rajeshrout97@gmail.com" },
]
keywords = ["terminal", "remote", "ai", "mcp", "cursor", "ollama", "macos"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: MacOS",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: System :: Networking",
    "Topic :: Software Development :: Libraries",
]
dependencies = [
    "websockets>=12.0",
    "mcp>=1.0",
    "click>=8.0",
    "pyyaml>=6.0",
    "cryptography>=42.0",
    "keyring>=25.0",
    "rich>=13.0",
    "pydantic>=2.0",
    "aiohttp>=3.9",
    "qrcode>=7.0",
    "zeroconf>=0.131.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.3",
]

[project.scripts]
tbridge = "terminal_bridge.cli:main"

[project.urls]
Homepage = "https://github.com/rajeshrout97/terminal-bridge"
Documentation = "https://github.com/rajeshrout97/terminal-bridge#readme"
Repository = "https://github.com/rajeshrout97/terminal-bridge"
Issues = "https://github.com/rajeshrout97/terminal-bridge/issues"

[tool.hatch.version]
path = "src/terminal_bridge/version.py"

[tool.hatch.build.targets.wheel]
packages = ["src/terminal_bridge"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

